refactor(client): move ResultRow class in specific file
