+++ client/src/Main/Monitor/ResultRow/index.js
@@ -1,0 +1,3 @@
+import ResultRow from './view';
+
+export default ResultRow;
